requirements.txt
opencv-python>=4.8.0
pyautogui>=0.9.54
pydirectinput>=1.0.4
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
mss>=10.0.0
